backend/requirements.txt
Django>=5.0,<5.3
djangorestframework>=3.15,<3.16
pandas>=2.2,<2.3
matplotlib>=3.8,<3.9
requests>=2.31,<3.0
celery>=5.3,<6.0
redis>=5.0,<6.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
tzdata>=2024.1
django-cors-headers>=4.3,<5.0
